Reworked structure of files and added basic links
Also added a configuration object

Changes were too intertwined to spend the time breaking them apart
