+++ lib/json_api.rb
@@ -4,3 +4,4 @@ end
 require_relative 'json_api/dsl'
 require_relative 'json_api/resource'
 require_relative 'json_api/exceptions'
+require_relative 'json_api/configuration'
