+++ spec/spec_helper.rb
@@ -2,7 +2,7 @@ require 'active_support'
 require 'active_support/core_ext'
 require 'pry-byebug'
 require_relative '../lib/json_api'
-require_relative 'support/resource_object'
+require_relative 'support/resource_objects'
 require_relative 'support/json_api_matchers'
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
